+++ HISTORY.md
@@ -1,9 +1,10 @@
 2.0.15
 -----------------
 - Unified behavior of listCollections results so 3.0 and pre 3.0 return same type of results.
-- Bumped mongodb-core to 1.1.10 to support per document tranforms in cursors.
+- Bumped mongodb-core to 1.1.11 to support per document tranforms in cursors as well as relaxing the setName requirement.
 - NODE-360 Aggregation cursor and command correctly passing down the maxTimeMS property.
 - Added ~1.0 mongodb-tools module for test running.
+- Remove the required setName for replicaset connections, if not set it will pick the first setName returned.
 
 2.0.14 2015-01-21
 -----------------